refactor: Winners 객체 생성
